+++ NEWS.md
@@ -1,5 +1,9 @@
 # targets 0.2.0.9000
 
+## New features
+
+* Implement an exponential backoff algorithm for polling the priority queue in `tar_make_clustermq()` and `tar_make_future()`. Add a `backoff` option in `tar_option_set()` to set the maximum upper bound (seconds) for the polling interval (#333).
+
 ## Enhancements
 
 * Use `rlang::check_installed()` inside `assert_package()` (#331, @malcolmbarrett).